refactor(desktop): decompose sendturn into lifecycle helpers
